refactor: move zip folder to relevant folder
