+++ 03-file-formats/scripts/lmdb/zip_folder.py
